requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
flet==0.25.2
flet-desktop==0.25.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
json5==0.10.0
lxml==5.3.0
multidict==6.1.0
numpy==2.2.0
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
repath==0.9.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
